Changed the SignalType enum to avoid failable initialization. If the raw value of the signal is not recognized, return .unknown as default value.

+++ Sources/POSIX/Sigaction.swift
@@ -80,6 +80,26 @@ public enum SignalType : Int32 {
     case usr1   = 30
     /// User-defined signal 2
     case usr2   = 31
+    /// Unknown or not supported signal
+    case unknown
+    
+    /// Initializer that avoids failable initialization. If the raw value is not recognized,
+    /// uses .unknown as default value.
+    ///
+    /// - Parameter rawValue: Signal number
+    public init(rawValue: Int32) {
+        switch rawValue {
+        case  1: self = .hup;   case  2: self = .int;       case  3: self = .quit;  case  4: self = .ill;
+        case  5: self = .trap   case  6: self = .abrt;      case  7: self = .poll;  case  8: self = .fpe;
+        case  9: self = .kill;  case 10: self = .bus        case 11: self = .segv;  case 12: self = .sys;
+        case 13: self = .pipe;  case 14: self = .alrm;      case 15: self = .term   case 16: self = .urg;
+        case 17: self = .stop;  case 18: self = .stp;       case 19: self = .cont;  case 20: self = .chld
+        case 21: self = .ttin;  case 22: self = .ttou;      case 23: self = .io;    case 24: self = .xcpu;
+        case 25: self = .xfsz   case 26: self = .vtalrm;    case 27: self = .prof;  case 28: self = .winch;
+        case 29: self = .info;  case 30: self = .usr1       case 31: self = .usr2
+        default: self = .unknown
+        }
+    }
 }
 
 /// Defines the kind of action to take when a signal is delivered
@@ -93,8 +113,10 @@ public enum SignalAction : Int32 {
 }
 
 public protocol SignalHandlerDelegate {
-    mutating func handleSignal(signal: SignalType?)
+    mutating func handleSignal(signal: SignalType)
 }
+
+public typealias SignalResponder = (SignalType?)->Void
 
 /// This class encapsulates all necessary functionality to handle signals sent by a given operating system.
 /// - remarks: 
@@ -112,6 +134,9 @@ public struct Signal {
     ///     - signal: Signal to add treatment to
     ///     - action: What to do when the signal is delivered.
     public static func setTrap(signal: SignalType, action: SignalAction) {
+        guard signal != .unknown else {
+            return
+        }
         CPOSIXInstallSignalHandler(signal.rawValue, action.rawValue) { (signal) in
             Signal.handleSignal(signal: signal)
         }
@@ -123,6 +148,9 @@ public struct Signal {
     ///     - pid: The pid to send the signal to. Defaults to current process
     ///     - signal: What signal to send.
     public static func killPid(pid: pid_t = getpid(), signal: SignalType) {
+        guard signal != .unknown else {
+            return
+        }
         kill(pid, signal.rawValue)
     }
